requirements.txt
yfinance
pyzmq